% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arknpi-package.R
\docType{package}
\name{arknpi-package}
\alias{arknpi}
\alias{arknpi-package}
\title{arknpi: Archive: NPPES NPI Registry}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/arknpi}
  \item \url{https://andrewallenbruce.github.io/arknpi/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/arknpi/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
