% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_date}
\alias{as_date}
\title{Format a Date}
\usage{
as_date(x, fmt = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{character string}

\item{fmt}{string format, e.g. '\%m/\%d/\%Y'}
}
\value{
date string
}
\description{
Format a Date
}
\examples{
as_date("12/31/2021")

}
