% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entity_recode}
\alias{entity_recode}
\title{Recode NPI entity codes}
\usage{
entity_recode(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector
}
\description{
Recode NPI entity codes
}
\examples{
entity_recode(c("1", "2", "3"))

}
