% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_phone}
\alias{make_phone}
\title{Format a 10-digit phone number}
\usage{
make_phone(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector
}
\description{
Format a 10-digit phone number
}
\examples{
make_phone("9122474701")

make_phone("9122474701  9122474701")

}
