% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_purpose}
\alias{make_purpose}
\title{Make an Address Purpose Column}
\usage{
make_purpose(x, y)
}
\arguments{
\item{x}{Mailing address column}

\item{y}{Practice address column}
}
\value{
character string:
\itemize{
\item If x is not \code{NA} and y is, \strong{\code{M}}
\item If x is \code{NA} and y is not, \strong{\code{P}}
\item Otherwise, \strong{\code{MP}}
}
}
\description{
Make an Address Purpose Column
}
\examples{
make_purpose(c("1", NA, NA), c(NA, "2", NA))

}
