% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_zip}
\alias{make_zip}
\title{Format a 9-digit zip code}
\usage{
make_zip(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector
}
\description{
Format a 9-digit zip code
}
\examples{
make_zip(c("31605", "316051234", NA))

}
