% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{other_id_type}
\alias{other_id_type}
\title{NPPES Other Identifier Type}
\usage{
other_id_type(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of \code{otid} types}
}
\value{
\verb{<chr>} vector of \code{otid} descriptions
}
\description{
NPPES Other Identifier Type
}
\examples{
data.table::data.table(
  otid_type = as.character(1:8),
  otid_desc = other_id_type(as.character(1:8)))

}
\seealso{
Other NPPES Decoders: 
\code{\link{other_name_type}()}
}
\concept{NPPES Decoders}
