% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{other_name_type}
\alias{other_name_type}
\title{NPPES Other Name Type}
\usage{
other_name_type(x)
}
\arguments{
\item{x}{\verb{<chr>} vector; \code{other_last_type} or \code{other_org_type}}
}
\value{
\verb{<chr>} vector of \verb{other_*_type} descriptions
}
\description{
NPPES Other Name Type
}
\examples{
dplyr::tibble(
 oname_type = as.character(1:5),
 oname_desc = other_name_type(oname_type))

}
\seealso{
Other NPPES Decoders: 
\code{\link{other_id_type}()}
}
\concept{NPPES Decoders}
