% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phone_eq_10}
\alias{phone_eq_10}
\title{Format a 10-digit phone number}
\usage{
phone_eq_10(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector
}
\description{
Format a 10-digit phone number
}
\examples{
phone_eq_10(c("9122474701", NA))

}
