% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_periods}
\alias{remove_periods}
\title{Remove periods}
\usage{
remove_periods(x)
}
\arguments{
\item{x}{character string}
}
\value{
character string
}
\description{
Remove periods
}
\examples{
remove_periods("M.D..")

}
