% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap}
\alias{wrap}
\title{Wrap a string}
\usage{
wrap(x, left, right)
}
\arguments{
\item{x}{character string}

\item{left}{character string}

\item{right}{character string}
}
\value{
character string
}
\description{
Wrap a string
}
\examples{
wrap("912", "(", ")")

}
