% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_case}
\alias{make_case}
\title{Make a case statement}
\usage{
make_case(x, y)
}
\arguments{
\item{x}{A character vector}

\item{y}{A character vector}
}
\value{
A character vector
}
\description{
Make a case statement
}
\examples{
make_case(c("1", "1", "3", NA), c("1", "2", NA, "4"))

}
